requirements.txt
asgiref==3.2.10
Django==3.1.2
djangorestframework==3.12.1
gunicorn==20.0.4
pika==1.1.0
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1
